requirements.txt

botcity-maestro-sdk>=0.5.1,<1.0
dotenv
instagrapi
requests
# Uncomment to use botcity-utils
# botcity-utils>=0.1.2

# Uncomment to use botcity-crawler-plugin
# botcity-crawler-plugin

# Uncomment to use Pandas
# pandas
